requirements.txt
flask==3.1.0
flask-session==0.8.0
argon2-cffi==23.1.0
bleach==6.2.0
flasgger==0.9.7b2